requirements.txt
tesseract-ocr
Pillow
pytesseract
transformers
streamlit
streamlit-nested-layout
torch==1.11.0
sentencepiece